move ghosttext component to textfield
